+++ src/pages/index.js
@@ -5,10 +5,9 @@ import { HowItWorks } from './home/_how-it-works'
 import { Markets } from './home/_markets'
 import { WhyDeriv } from './home/_why-deriv'
 import { WhyDerivMobile } from './home/_why-deriv-mobile'
-import Payment from './home/_payment-methods'
+import PaymentMethods from './home/_payment-methods'
 import SEO from 'components/containers/seo'
 import Layout from 'components/layout/layout'
-import PaymentMethods from 'components/elements/payment-methods'
 import { localize, WithIntl } from 'components/localization'
 import { Divider } from 'components/elements/divider'
 import Show from 'components/containers/show'
@@ -22,7 +21,6 @@ const Home = () => (
             )}
         />
         <Hero />
-        <Payment />
         <Show.Mobile>
             <WhyDerivMobile />
             <Dtrader />
